requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
tensorflow
keras
opencv-python
numpy
pandas
pytesseract
pypdf2
scikit-learn
matplotlib
seaborn
reportlab
python-dotenv
datasets
huggingface_hub
pillow
google-generativeai
flask-socketio
pydicom
python-socketio
eventlet
